Fix issue #8.
In the Python version of this code, there's a wrap_ord function that converts
characters into numbers. In lib/rchardet/jpcntx.rb there's no mention of this
function and we're trying to simply take numbers away from chars, which does
not work. Fixed this so that we're subtracting from the character's ordinal
number.